% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPI_window_testbed.R
\name{NPI_window_testbed}
\alias{NPI_window_testbed}
\title{Run the Network Parameter Outlier (NPO) algorithm.}
\usage{
NPI_window_testbed(
  path,
  taskName,
  institution,
  lab,
  experiment,
  subject,
  signalType,
  centerTime,
  iterationType,
  hostname,
  dbName,
  db_user,
  db_password,
  file_password,
  range,
  minimum_sampling_frequency,
  waveform_mask,
  computation_mask,
  database_update_limit,
  filter_detect_lowF,
  filter_detect_highF,
  filter_keep_lowF,
  filter_keep_highF,
  correlationWindow,
  CCthreshold,
  EDthreshold,
  blackout
)
}
\description{
Run the Network Parameter Outlier (NPO) algorithm.
}
\examples{
\dontrun{
}
}
