% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPO_window_testbed.R
\name{NPO_window_testbed}
\alias{NPO_window_testbed}
\title{Run the Network Parameter Outlier (NPO) algorithm.}
\usage{
NPO_window_testbed(...)
}
\description{
Run the Network Parameter Outlier (NPO) algorithm.
}
\examples{
\dontrun{
  First time the function is called:
  NPO_window(dbName='NSME_halo',path='NPO/Analysis/NPO/tests/testData/Halo/11',data='rodentMSO',institution='Yale',lab='NSME',experiment='Halo10sec_10x',subject=11,signalType='AP',centerTime=0,iterationType='directory',range=c(-3000,2000), '--restart' )

... or ...

  NPO_window(dbName='NV', path='/Users/markrbower/Dropbox/Documents/Concepts/2019_11_19_NetworkParameterOutlier/mef2', data='NV_23_002', institution='Yale', lab='NSME', experiment='baseline', subject='23_002', signalType='IIS', centerTime=0, iterationType='directory', range=c(-3000,2000), '--restart' )

  NPO_window(service='NV', path='/Users/markrbower/Documents/Data/NV/NVC1001_24_005_2', taskName='preprocessing', institution='Yale', lab='NSME', experiment='NeuroVista', subject='24_005', signalType='IIS', centerTime=0, iterationType='directory', range=c(-3000,2000), '--restart' )

  Subsequent calls:
  NPO()
}
}
